Move `ServerRequest` class to `React\Http\Message\ServerRequest`
