+++ README.md
@@ -70,6 +70,7 @@ multiple concurrent HTTP requests without blocking.
         * [withResponseBuffer()](#withresponsebuffer)
     * [React\Http\Message](#reacthttpmessage)
         * [Response](#response)
+        * [ServerRequest](#serverrequest)
     * [React\Http\Middleware](#reacthttpmiddleware)
         * [StreamingRequestMiddleware](#streamingrequestmiddleware)
         * [LimitConcurrentRequestsMiddleware](#limitconcurrentrequestsmiddleware)
@@ -2355,6 +2356,26 @@ which in turn extends the
   class. The only difference is that this class will accept implemenations
   of ReactPHPs `ReadableStreamInterface` for the `$body` argument. This base
   class is considered an implementation detail that may change in the future.
+
+#### ServerRequest
+
+The `ServerRequest` class can be used to
+respresent an incoming server request message.
+
+This class implements the
+[PSR-7 `ServerRequestInterface`](https://www.php-fig.org/psr/psr-7/#321-psrhttpmessageserverrequestinterface)
+which extends the
+[PSR-7 `RequestInterface`](https://www.php-fig.org/psr/psr-7/#32-psrhttpmessagerequestinterface)
+which in turn extends the
+[PSR-7 `MessageInterface`](https://www.php-fig.org/psr/psr-7/#31-psrhttpmessagemessageinterface).
+
+This is mostly used internally to represent each incoming request message.
+Likewise, you can also use this class in test cases to test how your web
+application reacts to certain HTTP requests.
+
+> Internally, this implementation builds on top of an existing outgoing
+  request message and only adds required server methods. This base class is
+  considered an implementation detail that may change in the future.
 
 ### React\Http\Middleware
