+++ src/Io/RequestHeaderParser.php
@@ -4,6 +4,7 @@ namespace React\Http\Io;
 
 use Evenement\EventEmitter;
 use Psr\Http\Message\ServerRequestInterface;
+use React\Http\Message\ServerRequest;
 use React\Socket\ConnectionInterface;
 use Exception;
