+++ src/Message/ServerRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace React\Http\Io;
+namespace React\Http\Message;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
@@ -8,18 +8,24 @@ use Psr\Http\Message\UriInterface;
 use RingCentral\Psr7\Request;
 
 /**
- * [Internal] Implementation of the PSR-7 `ServerRequestInterface`
+ * Respresents an incoming server request message.
  *
- * This is used internally to represent each incoming request message.
+ * This class implements the
+ * [PSR-7 `ServerRequestInterface`](https://www.php-fig.org/psr/psr-7/#321-psrhttpmessageserverrequestinterface)
+ * which extends the
+ * [PSR-7 `RequestInterface`](https://www.php-fig.org/psr/psr-7/#32-psrhttpmessagerequestinterface)
+ * which in turn extends the
+ * [PSR-7 `MessageInterface`](https://www.php-fig.org/psr/psr-7/#31-psrhttpmessagemessageinterface).
  *
- * This implementation builds on top of an existing outgoing request message and
- * only adds required server methods.
+ * This is mostly used internally to represent each incoming request message.
+ * Likewise, you can also use this class in test cases to test how your web
+ * application reacts to certain HTTP requests.
  *
- * Note that this is an internal class only and nothing you should usually care
- * about. See the `ServerRequestInterface` for more details.
+ * > Internally, this implementation builds on top of an existing outgoing
+ *   request message and only adds required server methods. This base class is
+ *   considered an implementation detail that may change in the future.
  *
  * @see ServerRequestInterface
- * @internal
  */
 class ServerRequest extends Request implements ServerRequestInterface
 {
