+++ tests/Middleware/RequestBodyBufferMiddlewareTest.php
@@ -6,9 +6,9 @@ use Clue\React\Block;
 use Psr\Http\Message\ServerRequestInterface;
 use React\EventLoop\Factory;
 use React\Http\Io\HttpBodyStream;
-use React\Http\Io\ServerRequest;
-use React\Http\Middleware\RequestBodyBufferMiddleware;
 use React\Http\Message\Response;
+use React\Http\Message\ServerRequest;
+use React\Http\Middleware\RequestBodyBufferMiddleware;
 use React\Stream\ThroughStream;
 use React\Tests\Http\TestCase;
 use RingCentral\Psr7\BufferStream;
